+++ examples/rtu-server-address.rs
@@ -51,12 +51,20 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Give the server some time for stating up
     thread::sleep(Duration::from_secs(1));
 
-    println!("Connecting client...");
+    println!("CLIENT: Connecting client...");
     let client_serial = tokio_serial::SerialStream::open(&builder).unwrap();
     let mut ctx = rtu::attach_slave(client_serial, slave);
-    println!("Reading input registers...");
+    println!("CLIENT: Reading input registers...");
     let rsp = ctx.read_input_registers(0x00, 7).await?;
-    println!("The result is '{rsp:#x?}'"); // The result is '[0x0,0x0,0x77,0x0,0x0,0x0,0x0,]'
+    println!("CLIENT: The result is '{rsp:#x?}'");
+    assert_eq!(rsp, Ok(vec![0x0, 0x0, 0x77, 0x0, 0x0, 0x0, 0x0]));
+
+    println!("CLIENT: Reading with illegal function... (should return IllegalFunction)");
+    let response = ctx.read_holding_registers(0x100, 1).await.unwrap();
+    println!("CLIENT: The result is '{response:?}'");
+    assert_eq!(response, Err(Exception::IllegalFunction));
+
+    println!("CLIENT: Done.");
 
     Ok(())
 }