+++ src/client/tcp.rs
@@ -3,7 +3,7 @@
 
 //! TCP client connections
 
-use std::{fmt, io::Error, net::SocketAddr};
+use std::{fmt, io, net::SocketAddr};
 
 use tokio::{
     io::{AsyncRead, AsyncWrite},
@@ -13,14 +13,14 @@ use tokio::{
 use super::*;
 
 /// Establish a direct connection to a Modbus TCP coupler.
-pub async fn connect(socket_addr: SocketAddr) -> Result<Context, Error> {
+pub async fn connect(socket_addr: SocketAddr) -> io::Result<Context> {
     connect_slave(socket_addr, Slave::tcp_device()).await
 }
 
 /// Connect to a physical, broadcast, or custom Modbus device,
 /// probably through a Modbus TCP gateway that is forwarding
 /// messages to/from the corresponding slave device.
-pub async fn connect_slave(socket_addr: SocketAddr, slave: Slave) -> Result<Context, Error> {
+pub async fn connect_slave(socket_addr: SocketAddr, slave: Slave) -> io::Result<Context> {
     let transport = TcpStream::connect(socket_addr).await?;
     let context = attach_slave(transport, slave);
     Ok(context)